Se ha decido deprecar el uso de Repository. Es por eso que se decidió extraer de commons-repository-* aquellas clases que sirven más allá de Repository (auditoria, interceptores, ...); para facilitar el aisalamiento. Se crearon los proyectos commons-persistence-api; y commons-persistence-util a donde se mueven varias clases. commons-repository-api depende de commons-persistence-api y commons-repository-springhbm de commons-persistence-util por lo que las aplicaciones que usan repository van a ver pocos combios (los paquetes se mantienen). Todos los cambios están pensados para requerir la menor cantidad de cambios para que siga funcionando. Aquellos proyectos que dependian de commons-repository-springhbm debido a los interceptores; pero que no usaban la clase Repository es recomendable que cambien sus dependencias hacia commons-persistence-util. En commons-persistence-api quedaron las interfaces: Nameable, CreationAuditable, LogicallyDeleteable, Modifiable, ModificationAuditable. Todas ellas antes extendían de Persistible que tenía un id Long. Esta extensión dejó de existir; para no forzar herencias; y para que queden como interfaces de marca; por lo que hay que agregar en todas las entidades que implementaban alguna de estas interfaces, deben agregar a la lista de implements a Persistible (si es que la usaban) (que ha quedado en commons-repository-api).

+++ spring/src/main/java/ar/com/zauber/commons/spring/web/controllers/AbstractRestishController.java
@@ -74,13 +74,15 @@ public abstract class AbstractRestishController extends AbstractController {
         } else if(patharray.length == firstIndex) {
             ret = list(request, response);
         } else if(patharray.length == firstIndex + 1) {
-            final Long id = seoStrategy.getIdFromSEOFriendly(patharray[firstIndex]);
+            final Long id = (Long) seoStrategy.getIdFromSEOFriendly(
+                    patharray[firstIndex]);
             if(id == null) {
                 throw new NoSuchEntityException(patharray[firstIndex]);
             }
             ret = entity(id, request, response);
         } else if(patharray.length == firstIndex + 2) {
-            final Long id = seoStrategy.getIdFromSEOFriendly(patharray[firstIndex]);
+            final Long id = (Long) seoStrategy.getIdFromSEOFriendly(
+                    patharray[firstIndex]);
             if(id == null) {
                 throw new NoSuchEntityException(patharray[firstIndex]);
             }
